mira changes

+++ pidx/data_handle/PIDX_memory_layout_data_structs.h
@@ -92,7 +92,7 @@ struct PIDX_HZ_Agg_buffer_struct
 
   int num_idx_blocks;
   uint64_t buffer_size;                                 ///< Aggregator buffer size
-  uint64_t compressed_buffer_size;                      ///< Aggregator buffer size after compression
+  unsigned long long compressed_buffer_size;                      ///< Aggregator buffer size after compression
   uint64_t *compressed_block_size;                      ///< Compressed size of each block in the aggregator
   unsigned char* buffer;                                ///< The actual aggregator buffer
 };
